Refactoring: removed ProjectilePool class
Refactoring: removed ProjectilePool class
