--- Assets/Scripts/Gameplay/Projectiles/ProjectilePool.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 1a1db8a89abe411492f0a12f96d52240
-timeCreated: 1695218635
