+++ Assets/Scripts/Gameplay/Weapons/IWeapon.cs
@@ -1,4 +1,5 @@
 using System;
+using Gameplay.Projectiles;
 using Gameplay.Services.GameTime;
 
 namespace Gameplay.Weapons
@@ -8,6 +9,7 @@ namespace Gameplay.Weapons
         bool Shot();
         void Dispose();
         void Reload();
+        void RemoveProjectile(Projectile projectile);
 
         public class Fake : IWeapon
         {
@@ -20,6 +22,7 @@ namespace Gameplay.Weapons
             public bool Shot() => false;
             public void Dispose(){}
             public void Reload(){}
+            public void RemoveProjectile(Projectile projectile){}
         }
     }
 }
