+++ Assets/Scripts/Gameplay/Weapons/Installer/WeaponInstaller.cs
@@ -7,7 +7,6 @@ namespace Gameplay.Weapons
     public class WeaponInstaller : MonoInstaller
     {
         [SerializeField] private Weapon _weapon;
-        [SerializeField] private ProjectilePool _projectilePool;
 
         [Inject] private IWeaponSettings _settings;
         [Inject] private IWeaponUser _weaponUser;
@@ -17,7 +16,6 @@ namespace Gameplay.Weapons
             BindInjectedParameters();
             BindWeaponSelf();
             BindProjectile();
-            BindProjectilePool();
                 
             BindSound();
         }
@@ -49,15 +47,7 @@ namespace Gameplay.Weapons
                 .NonLazy();
         }
 
-        private void BindProjectilePool()
-        {
-            Container.Bind<ProjectilePool>()
-                .FromInstance(_projectilePool)
-                .AsCached()
-                .NonLazy();
-        }
-
-        private void BindProjectile()
+       private void BindProjectile()
         {
             Container.BindFactory<FlyInfo, IAttackInfo, Projectile, Projectile.Factory>()
                 .FromMonoPoolableMemoryPool(pool => pool
